修正S_RangeSkill 目標數超過128報錯的可能 隨機函數修正
git-svn-id: http://www.castiello-tsai.org:8523/svn/L1JTW@163 96aee97a-5c11-ed42-8518-d7f33e96f5f9

+++ src/l1j/server/server/serverpackets/S_RangeSkill.java
@@ -67,7 +67,7 @@ public class S_RangeSkill extends ServerBasePacket {
 		writeC(type); // 0:範圍 6:遠距離 8:範圍&遠距離
 		writeH(0);
 		writeH(target.length);
-		for (byte i = 0; i < target.length; i++) {
+		for (short i = 0; i < target.length; i++) {
 			writeD(target[i].getId());
 			writeC(0x20); // 0:ダメージモーションあり 0以外:なし
 		}
